03-Curso_2/09-others/patterns/requirements.txt
aws-cdk-lib>=2.146.0
constructs>=10.0.0,<11.0.0
aws-solutions-constructs.core>=2.60.0
aws_solutions_constructs.aws_s3_lambda>=2.60.0